Remove Input.v2
