+++ src/components/Input/Input.tsx
@@ -1,10 +1,10 @@
-import React, { PropsWithChildren } from "react";
+import React, { PropsWithChildren, useMemo, useState } from "react";
 import { TextInput, TextInputProps } from "react-native";
 
 import { ShortcutProps, defaultShortcuts } from "@styles/shortcuts";
-import { HStack } from "../HStack/HStack";
+import { HStack } from "@components/HStack/HStack";
 import { defaults, variants } from "./styles";
-import { InputLeftElement, InputRightElement } from "./components";
+import { InputLeftElement, InputRightElement, FloatingLabel } from "./components";
 
 interface Props extends PropsWithChildren, TextInputProps, ShortcutProps {
   isDisabled?: boolean;
@@ -15,20 +15,44 @@ interface Props extends PropsWithChildren, TextInputProps, ShortcutProps {
 }
 
 export function Input(props: Props) {
-  const { onChangeText, isDisabled, inputLeftElement, inputRightElement, style, ...restProps } =
-    props;
+  const {
+    onChangeText,
+    isDisabled,
+    inputLeftElement,
+    inputRightElement,
+    style,
+    variant = "standard",
+    placeholder: label,
+    value,
+    ...restProps
+  } = props;
+
+  const [isFocused, setIsFocused] = useState(false);
+  const mustFloatLabel = useMemo(() => Boolean(isFocused || value), [isFocused, value]);
+
+  const handleFocus = () => {
+    setIsFocused(true);
+  };
+
+  const handleBlur = () => {
+    setIsFocused(false);
+  };
 
   return (
     <HStack style={[defaults().wrapper, defaultShortcuts(props)]} alignItems="center">
+      {label && <FloatingLabel label={label} isFloat={mustFloatLabel} />}
       <InputLeftElement shown={Boolean(inputLeftElement)}>{inputLeftElement}</InputLeftElement>
       <TextInput
+        placeholderTextColor="white"
         onChangeText={text => !isDisabled && onChangeText?.(text)}
         style={[
           defaults({ hasLeftElement: Boolean(inputLeftElement) }).input,
-          variants[props.variant!].input,
+          variants[variant].input,
           style,
         ]}
         {...restProps}
+        onFocus={handleFocus}
+        onBlur={handleBlur}
       />
       <InputRightElement shown={Boolean(inputRightElement)}>{inputRightElement}</InputRightElement>
     </HStack>
@@ -38,5 +62,5 @@ export function Input(props: Props) {
 Input.defaultProps = {
   isDisabled: false,
   isInvalid: false,
-  variant: "filled",
+  variant: "standard",
 };
