+++ src/components/Input/PasswordInput.tsx
@@ -1,7 +1,7 @@
 import React, { PropsWithChildren } from "react";
 import { TextInputProps, TouchableOpacity } from "react-native";
 import Icon from "react-native-vector-icons/Ionicons";
-import { Input } from "./Input.v2";
+import { Input } from "./Input";
 
 interface Props extends PropsWithChildren, TextInputProps {
   isDisabled?: boolean;
